+++ examples/redux-thunk/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/Jepria/jfront-core/compare/jfront-core-redux-thunk@0.1.1...jfront-core-redux-thunk@0.1.2) (2021-01-21)
+
+
+### Bug Fixes
+
+* change name searchTemplate -> searchRequest in SearchState. Added pageSize, pageNumber ([be56b36](https://github.com/Jepria/jfront-core/commit/be56b36a33847dd947fc6e954b38bf72567a0753))
+
+
+
+
+
 ## [0.1.1](https://github.com/Jepria/jfront-core/compare/jfront-core-redux-thunk@0.1.0...jfront-core-redux-thunk@0.1.1) (2021-01-18)
 
 **Note:** Version bump only for package jfront-core-redux-thunk